feat: integrates lemmatizer into word search
Integrates lemmatizer from Stanza (StanfordNLP) instead of stemmer, which guarantees that the processed word
is a word from a dictionary. Tradeoff for root word discovery is accuracy for speed since stemmer is very lightweight in comparison.
We were able to run the lemmatizer on a laptop that doesn't have a dedicated GPU however, so it doesn't seem that heavy to run.